+++ src/pages/Admin/ApplicantDetail/index.tsx
@@ -3,8 +3,8 @@ import { Window } from "$models/Window";
 import { useParams } from "react-router-dom";
 import styles from "./styles.module.scss";
 
-import { ApplicantDetailInfo } from "../Home/Dashboard/TaskDetail/Applicant/ApplicantDetailInfo";
-import { ApplicantDetailContent } from "../Home/Dashboard/TaskDetail/Applicant/ApplicantDetailContent";
+import { ApplicantDetailInfo } from "../components/Applicant/ApplicantDetailInfo";
+import { ApplicantDetailContent } from "../components/Applicant/ApplicantDetailContent";
 import { Window as WindowComponent } from "$components/Window";
 
 export const ApplicantDetail: FunctionComponent = () => {
