+++ src/pages/Admin/Home/Dashboard/TaskDetail/JobApplication/DetailInfo/component.tsx
@@ -2,7 +2,7 @@ import React, { FunctionComponent } from "react";
 import { ApprovalStatus } from "$interfaces/ApprovalStatus";
 import { IJobApplication } from "$interfaces/JobApplication";
 import { DetailInfo } from "../../../../../components/DetailInfo";
-import { UserDetails } from "../../Applicant/ApplicantDetailInfo/UserDetails";
+import { UserDetails } from "../../../../../components/Applicant/ApplicantDetailInfo/UserDetails";
 import { MainTitle } from "./MainTitle";
 
 export const JobApplicationDetailInfo: FunctionComponent<IComponentProps> = ({
