+++ src/pages/Admin/Home/Dashboard/TaskDetail/component.tsx
@@ -1,10 +1,10 @@
 import React, { FunctionComponent, useRef } from "react";
 import { CompanyDetailContent } from "./Company/CompanyDetailContent";
-import { ApplicantDetailContent } from "./Applicant/ApplicantDetailContent";
+import { ApplicantDetailContent } from "../../../components/Applicant/ApplicantDetailContent";
 import { OfferDetailContent } from "./Offer/OfferDetailContent";
 import { JobApplicationDetailContent } from "./JobApplication/DetailContent";
 import { CompanyDetailInfo } from "./Company/CompanyDetailInfo";
-import { ApplicantDetailInfo } from "./Applicant/ApplicantDetailInfo";
+import { ApplicantDetailInfo } from "../../../components/Applicant/ApplicantDetailInfo";
 import { OfferDetailInfo } from "./Offer/OfferDetailInfo";
 import { JobApplicationDetailInfo } from "./JobApplication/DetailInfo";
 import { EmptyDetail } from "./EmptyDetail";
