+++ src/pages/Admin/components/Applicant/ApplicantDetailContent/index.ts
