+++ src/pages/Admin/components/Applicant/ApplicantDetailInfo/MainTitle/index.ts
