+++ src/pages/Admin/components/Applicant/ApplicantDetailInfo/UserDetails/container.tsx
