+++ src/pages/Admin/components/Applicant/ApplicantDetailInfo/UserDetails/index.tsx
