+++ src/pages/Admin/components/Applicant/ApplicantDetailInfo/UserDetails/interfaces.ts
