+++ src/pages/Admin/components/Applicant/ApplicantDetailInfo/UserDetails/styles.module.scss
