+++ src/pages/Admin/components/Applicant/ApplicantDetailInfo/component.tsx
@@ -1,7 +1,7 @@
 import React, { FunctionComponent } from "react";
 import { ApprovalStatus } from "$interfaces/ApprovalStatus";
 import { IApplicant } from "$interfaces/Applicant";
-import { DetailInfo } from "../../../../../components/DetailInfo";
+import { DetailInfo } from "../../DetailInfo";
 import { UserDetails } from "./UserDetails";
 import { MainTitle } from "./MainTitle";
