+++ src/pages/Admin/components/Applicant/ApplicantDetailInfo/index.ts
